+++ docs/CHANGELOG.md
@@ -24,6 +24,7 @@ It features the following interaction:
     * ``tomorrow_test_count``: number of tests for the next day.
     * ``phonetic_reading``: phonetic reading of a text.
     * ``theme_path``: theming path for a file.
+    * ``texts_statistics``: various words statistics for each text.
   * On POST, ``action_type`` can be:
     * ``text`` of ``audio`` change the reading position for a text or it's audio.
     * ``simterms``: similar terms to a given term
@@ -72,6 +73,7 @@ the intendended page. Same goes for "New Text" on "Archived Texts" page.
   * ``inc/ajax_get_theme.php`` should be accessed through ``action_type=theme_path``. Argument ``filepath`` is now ``path``.
   * ``inc/ajax_save_setting.php`` should be accessed through ``action_type=save_setting``, same arguments.
   * ``inc/ajax_save_text_position.php`` should be accessed through ``action=reading_position``, arguments changed.
+  * ``inc/ajax_word_counts.php`` should be accessed through ``action=texts_statistics``, on GET. Argument ``id`` is now ``texts_id``.
 
 ## 2.8.1-fork (April 14 2023)